requirements.txt
streamlit
plotly
streamlit-autorefresh
pandas
pyyaml
rich
pyfiglet
